requirements.txt
# GMM Product Recommendation System — Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
scikit-learn>=1.2.0
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0
joblib>=1.2.0

# Configuration
PyYAML>=6.0

# AWS / SageMaker (optional — only needed for cloud deployment)
boto3>=1.26.0
sagemaker>=2.140.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Notebooks (optional)
jupyter>=1.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
